Works: Progress, Results, and reading user profile

+++ Domain/Test/Progress.hpp
@@ -21,7 +21,7 @@ class Progress
 	  bool read_profile();
 	  ~Progress();
 	  Result locateResult(string);
-	  double getAverageWPM();
+	  float getAverageWPM();
 	  void updateResults(string, Result);
 	  void print_results();
 };
